requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
python-docx
flask_mongoengine
websocket
websocket-client
vosk
# pudub
requests